fix: refresh lens token (#2555)
* refactor: code style

* refactor: code style

* fix: retry lens session

* refactor: lint

* refactor: profile id after refresh session
