+++ src/providers/lens/SessionHolder.ts
@@ -1,13 +1,12 @@
 import { LensClient as LensClientSDK } from '@lens-protocol/client';
-import { ZERO_ADDRESS } from '@masknet/web3-shared-evm';
 
-import { THIRTY_DAYS } from '@/constants/index.js';
 import {
     createLensSDK,
     LocalStorageProvider,
     removeLensCredentials,
     setLensCredentials,
 } from '@/helpers/createLensSDK.js';
+import { refreshLensSession } from '@/helpers/refreshLensSession.js';
 import { SessionHolder } from '@/providers/base/SessionHolder.js';
 import { LensSession } from '@/providers/lens/Session.js';
 
@@ -24,29 +23,9 @@ class LensSessionHolder extends SessionHolder<LensSession> {
     override async refreshSession() {
         this.assertSession();
 
-        const [accessTokenResult, refreshTokenResult, walletAddress] = await Promise.all([
-            lensSessionHolder.sdk.authentication.getAccessToken(),
-            lensSessionHolder.sdk.authentication.getRefreshToken(),
-            lensSessionHolder.sdk.authentication.getWalletAddress(),
-        ]);
+        const session = await refreshLensSession(this.sdk);
 
-        const profileId = lensSessionHolder.session?.profileId;
-        const accessToken = accessTokenResult.unwrap();
-        const refreshToken = refreshTokenResult.unwrap();
-        const now = Date.now();
-
-        const session =
-            accessToken && refreshToken && walletAddress && profileId
-                ? new LensSession(
-                      profileId,
-                      accessToken,
-                      now,
-                      now + THIRTY_DAYS,
-                      refreshToken,
-                      walletAddress ?? ZERO_ADDRESS,
-                  )
-                : null;
-        if (!session) throw new Error('Failed to refresh session');
+        // the sdk always maintain a latest session, thought no need to resume session here.
 
         return session;
     }
